new Integer(3) != new Integer(3)
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/apps/load-balancing-sims@10801 67a373e5-eb02-0410-a15c-ee090a768436

+++ phase7/sim/handlers/SskInsertHandler.java
@@ -35,7 +35,7 @@ public class SskInsertHandler extends MessageHandler implements EventTarget
 	private void checkCollision()
 	{
 		Integer old = node.fetchSsk (key);
-		if (old != null && old != data) {
+		if (old != null && old.intValue() != data) {
 			node.log (this + " collided");
 			if (prev == null) node.log (this + " collided locally");
 			else prev.sendMessage (new SskDataFound (id, old));
